@@ -1,22 +1,21 @@
 import React, { FunctionComponent } from "react";
+import { makeStyles, createStyles } from '@mui/styles';
 import {
   List,
-  ListItem,
+  ListItemButton,
   ListItemText,
   ListItemSecondaryAction,
   IconButton,
-  makeStyles,
-  type Theme,
-  createStyles,
   Chip,
   Typography,
   Grid,
+  Pagination,
   LinearProgress,
   Menu,
   MenuItem,
   Box,
-} from "@material-ui/core";
-import { MoreVert } from "@material-ui/icons";
+} from "@mui/material";
+import { MoreVert } from "@mui/icons-material";
 import {
   useBuildState,
   useBuildDispatch,
@@ -28,7 +27,6 @@ import { SkeletonList } from "../SkeletonList";
 import { formatDateTime } from "../../_helpers/format.helper";
 import { useSnackbar } from "notistack";
 import { TextValidator } from "react-material-ui-form-validator";
-import { Pagination } from "@material-ui/lab";
 import { Build } from "../../types";
 import { BaseModal } from "../BaseModal";
 import { buildsService } from "../../services";
@@ -36,7 +34,7 @@ import { useNavigate } from "react-router";
 import { buildTestRunLocation } from "../../_helpers/route.helpers";
 import { Tooltip } from "../Tooltip";
 
-const useStyles = makeStyles((theme: Theme) =>
+const useStyles = makeStyles(() =>
   createStyles({
     listContainer: {
       height: "100%",
@@ -46,6 +44,7 @@ const useStyles = makeStyles((theme: Theme) =>
       visibility: "hidden",
     },
     listItem: {
+      paddingRight: 48,
       "&:hover $listItemSecondaryAction": {
         visibility: "inherit",
       },
@@ -118,7 +117,7 @@ const BuildList: FunctionComponent = () => {
   }, [handlePaginationChange]);
 
   return (
-    <React.Fragment>
+    <>
       <Box height="91%" overflow="auto">
         <List>
           {loading ? (
@@ -128,12 +127,11 @@ const BuildList: FunctionComponent = () => {
           ) : (
             buildList.map((build) => (
               <React.Fragment key={build.id}>
-                <ListItem
+                <ListItemButton
                   selected={selectedBuild?.id === build.id}
-                  button
                   onClick={() => selectBuildCalback(build.id)}
                   classes={{
-                    container: classes.listItem,
+                    root: classes.listItem,
                   }}
                 >
                   <ListItemText
@@ -180,11 +178,12 @@ const BuildList: FunctionComponent = () => {
                   >
                     <IconButton
                       onClick={(event) => handleMenuClick(event, build)}
+                      size="large"
                     >
                       <MoreVert />
                     </IconButton>
                   </ListItemSecondaryAction>
-                </ListItem>
+                </ListItemButton>
                 {build.isRunning && <LinearProgress />}
               </React.Fragment>
             ))
@@ -212,7 +211,7 @@ const BuildList: FunctionComponent = () => {
               onClick={() => {
                 buildsService
                   .update(menuBuild.id, { isRunning: false })
-                  .then((b) =>
+                  .then(() =>
                     enqueueSnackbar(`${menuBuild.id} finished`, {
                       variant: "success",
                     }),
@@ -255,9 +254,9 @@ const BuildList: FunctionComponent = () => {
                 required
                 value={newCiBuildId}
                 inputProps={{
-                  onChange: (event: any) =>
-                    setNewCiBuildId((event.target as HTMLInputElement).value),
-                  "data-testId": "newCiBuildId",
+                  onChange: (event: React.ChangeEvent<HTMLInputElement>) =>
+                    setNewCiBuildId(event.target.value),
+                  "data-testid": "newCiBuildId",
                 }}
               />
             </React.Fragment>
@@ -267,7 +266,7 @@ const BuildList: FunctionComponent = () => {
               .update(menuBuild.id, {
                 ciBuildId: newCiBuildId,
               })
-              .then((b) => {
+              .then(() => {
                 toggleEditDialogOpen();
               })
               .catch((err) =>
@@ -292,7 +291,7 @@ const BuildList: FunctionComponent = () => {
           }
           onSubmit={() => {
             deleteBuild(buildDispatch, menuBuild.id)
-              .then((build) => {
+              .then(() => {
                 toggleDeleteDialogOpen();
                 enqueueSnackbar(
                   `Build #${menuBuild.number || menuBuild.id} deleted`,
@@ -316,7 +315,7 @@ const BuildList: FunctionComponent = () => {
           }}
         />
       )}
-    </React.Fragment>
+    </>
   );
 };
 
