@@ -1,11 +1,12 @@
-import { IconButton, makeStyles } from "@material-ui/core";
-import { NavigateNext, NavigateBefore } from "@material-ui/icons";
+import { IconButton } from "@mui/material";
+import { NavigateNext, NavigateBefore } from "@mui/icons-material";
 import React from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 import { TestRun } from "../../types";
 import { Tooltip } from "../Tooltip";
+import { makeStyles } from "@mui/styles";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(() => ({
   button: {
     width: 64,
     height: 64,
@@ -28,7 +29,7 @@ export const ArrowButtons: React.FunctionComponent<{
   const classes = useStyles();
 
   const navigateNext = () => {
-    if (selectedTestRunIndex + 1 < testRuns.length) {
+    if (testRuns.length > selectedTestRunIndex + 1) {
       const next = testRuns[selectedTestRunIndex + 1];
       handleNavigation(next.id);
     }
@@ -45,8 +46,8 @@ export const ArrowButtons: React.FunctionComponent<{
   useHotkeys("left", navigateBefore, [selectedTestRunIndex, handleNavigation]);
 
   return (
-    <React.Fragment>
-      {selectedTestRunIndex + 1 < testRuns.length && (
+    <>
+      {testRuns.length > selectedTestRunIndex + 1 && (
         <Tooltip title={"Hotkey: ArrowRight"}>
           <IconButton
             color="secondary"
@@ -55,6 +56,7 @@ export const ArrowButtons: React.FunctionComponent<{
               right: 0,
             }}
             onClick={navigateNext}
+            size="large"
           >
             <NavigateNext className={classes.icon} />
           </IconButton>
@@ -69,11 +71,12 @@ export const ArrowButtons: React.FunctionComponent<{
               left: 0,
             }}
             onClick={navigateBefore}
+            size="large"
           >
             <NavigateBefore className={classes.icon} />
           </IconButton>
         </Tooltip>
       )}
-    </React.Fragment>
+    </>
   );
 };
