@@ -1,12 +1,14 @@
-import React, { FunctionComponent, useCallback, useEffect } from "react";
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+import React, { FunctionComponent, useCallback } from "react";
 import { Stage, Layer, Image } from "react-konva";
 import Rectangle, { MIN_RECT_SIDE_PIXEL } from "../Rectangle";
 import { IgnoreArea } from "../../types/ignoreArea";
-import { Grid, makeStyles, CircularProgress } from "@material-ui/core";
+import { Grid, CircularProgress, type Theme } from "@mui/material";
 import { NoImagePlaceholder } from "./NoImageAvailable";
 import Konva from "konva";
+import { makeStyles } from "@mui/styles";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles((theme: Theme) => ({
   canvasContainer: {
     overflow: "auto",
     backgroundColor: "white",
@@ -23,9 +25,21 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
+export type ImageStateLoad = "loaded" | "loading" | "failed";
+
+type StageOffsetPosition = {
+  x: number;
+  y: number;
+};
+
+type StageStateArray = [
+  StageOffsetPosition,
+  React.Dispatch<React.SetStateAction<StageOffsetPosition>>,
+];
+
 interface IDrawArea {
   imageName: string | undefined;
-  imageState: [undefined | HTMLImageElement, "loaded" | "loading" | "failed"];
+  imageState: [undefined | HTMLImageElement, ImageStateLoad];
   tempIgnoreAreas: IgnoreArea[];
   ignoreAreas: IgnoreArea[];
   setIgnoreAreas: (ignoreAreas: IgnoreArea[]) => void;
@@ -33,25 +47,14 @@ interface IDrawArea {
   setSelectedRectId: (id: string) => void;
   deleteIgnoreArea?: (id: string) => void;
   onStageClick: (event: Konva.KonvaEventObject<MouseEvent>) => void;
-  stageOffsetState: [
-    { x: number; y: number },
-    React.Dispatch<React.SetStateAction<{ x: number; y: number }>>,
-  ];
-  stageInitPosState: [
-    { x: number; y: number },
-    React.Dispatch<React.SetStateAction<{ x: number; y: number }>>,
-  ];
-  stagePosState: [
-    { x: number; y: number },
-    React.Dispatch<React.SetStateAction<{ x: number; y: number }>>,
-  ];
-  stageScrollPosState: [
-    { x: number; y: number },
-    React.Dispatch<React.SetStateAction<{ x: number; y: number }>>,
-  ];
+  stageOffsetState: StageStateArray;
+  stageInitPosState: StageStateArray;
+  stagePosState: StageStateArray;
+  stageScrollPosState: StageStateArray;
   stageScaleState: [number, React.Dispatch<React.SetStateAction<number>>];
   drawModeState: [boolean, React.Dispatch<React.SetStateAction<boolean>>];
 }
+
 export const DrawArea: FunctionComponent<IDrawArea> = ({
   imageName,
   imageState,
@@ -83,16 +86,17 @@ export const DrawArea: FunctionComponent<IDrawArea> = ({
   const stageRef = React.useRef<Konva.Stage>(null);
   const scrollContainerRef = React.useRef<HTMLDivElement>(null);
 
-  const isDeleteKey = (e: KeyboardEvent) => {
-    return e.key === "Delete" || e.key === "Backspace";
+  const isDeleteKey = (event: KeyboardEvent) => {
+    return event.key === "Delete" || event.key === "Backspace";
   };
 
   const handleStageKeyDown = useCallback(
-    (e: any) => {
-      if (!deleteIgnoreArea || isModifierKeyPressed(e)) {
+    (event: KeyboardEvent) => {
+      if (!deleteIgnoreArea || isModifierKeyPressed(event)) {
         return;
       }
-      if (selectedRectId && isDeleteKey(e)) {
+
+      if (selectedRectId && isDeleteKey(event)) {
         deleteIgnoreArea(selectedRectId);
       }
     },
@@ -106,43 +110,49 @@ export const DrawArea: FunctionComponent<IDrawArea> = ({
     }
   }, [handleStageKeyDown]);
 
-  const isModifierKeyPressed = (e: any) => {
-    return e.altKey || e.ctrlKey || e.shiftKey;
+  const isModifierKeyPressed = (event: KeyboardEvent) => {
+    return event.altKey || event.ctrlKey || event.shiftKey;
   };
 
-  useEffect(() => {
+  React.useEffect(() => {
     scrollContainerRef.current?.scrollTo(stageScollPos.x, stageScollPos.y);
   }, [stageScollPos]);
 
-  const handleContentMousedown = (e: any) => {
+  const handleContentMousedown = (
+    event: Konva.KonvaEventObject<MouseEvent>,
+  ) => {
     if (stageRef.current) {
       const container = stageRef.current.container();
       container.tabIndex = 1;
     }
+
     if (!isDrawMode) {
       return;
     }
 
     const newArea: IgnoreArea = {
       id: Date.now().toString(),
-      x: e.evt.offsetX,
-      y: e.evt.offsetY,
+      x: event.evt.offsetX,
+      y: event.evt.offsetY,
       width: MIN_RECT_SIDE_PIXEL,
       height: MIN_RECT_SIDE_PIXEL,
     };
+
     setIgnoreAreas([...ignoreAreas, newArea]);
     setSelectedRectId(newArea.id);
     setIsDrawing(true);
   };
 
-  const handleContentMouseup = (e: any) => {
+  const handleContentMouseup = () => {
     if (isDrawing) {
       setIsDrawing(!isDrawing);
       setIsDrawMode(false);
     }
   };
 
-  const handleContentMouseMove = (e: any) => {
+  const handleContentMouseMove = (
+    event: Konva.KonvaEventObject<MouseEvent>,
+  ) => {
     if (!isDrawMode) return;
 
     if (isDrawing) {
@@ -151,192 +161,187 @@ export const DrawArea: FunctionComponent<IDrawArea> = ({
         if (i.id === selectedRectId) {
           // new width and height
           i.width = Math.max(
-            Math.round(e.evt.offsetX - i.x),
+            Math.round(event.evt.offsetX - i.x),
             MIN_RECT_SIDE_PIXEL,
           );
           i.height = Math.max(
-            Math.round(e.evt.offsetY - i.y),
+            Math.round(event.evt.offsetY - i.y),
             MIN_RECT_SIDE_PIXEL,
           );
           return i;
         }
+
         return i;
       });
+
       setIgnoreAreas(newShapesList);
     }
   };
 
   const imageLoaded = () => {
     return (
-      imageStatus === "loaded" && (
+      <div
+        className={classes.canvasContainer}
+        ref={scrollContainerRef}
+        onScroll={(event: React.SyntheticEvent<HTMLElement>) => {
+          setStageScrollPos({
+            x: (event.target as HTMLElement).scrollLeft,
+            y: (event.target as HTMLElement).scrollTop,
+          });
+        }}
+      >
         <div
-          className={classes.canvasContainer}
-          ref={scrollContainerRef}
-          onScroll={(event) => {
-            setStageScrollPos({
-              x: (event.target as HTMLElement).scrollLeft,
-              y: (event.target as HTMLElement).scrollTop,
+          style={{
+            height: image && image.height * stageScale,
+            width: image && image.width * stageScale,
+            transform: `translate3d(${stagePos.x}px, ${stagePos.y}px, 0px)`,
+          }}
+          onMouseMove={(event) => {
+            if (!isDrawMode && isDrag && !selectedRectId) {
+              event.preventDefault();
+              setStagePos({
+                x: event.clientX - stageInitPos.x,
+                y: event.clientY - stageInitPos.y,
+              });
+              setStageOffset(stagePos);
+            }
+          }}
+          onMouseUp={() => {
+            setIsDrag(false);
+            setStageInitPos(stagePos);
+          }}
+          onMouseLeave={() => {
+            setIsDrag(false);
+            setStageInitPos(stagePos);
+          }}
+          onMouseDown={(event) => {
+            setIsDrag(true);
+            setStageInitPos({
+              x: event.clientX - stageOffset.x,
+              y: event.clientY - stageOffset.y,
             });
           }}
+          onKeyDown={(event) => handleStageKeyDown(event)}
         >
-          <div
+          <Stage
+            ref={stageRef}
+            width={image?.width}
+            height={image?.height}
+            onMouseDown={onStageClick}
+            onWheel={(event: Konva.KonvaEventObject<WheelEvent>) => {
+              event.evt.preventDefault();
+              const scaleBy = 1.04;
+              const newScale =
+                event.evt.deltaY < 0
+                  ? stageScale * scaleBy
+                  : stageScale / scaleBy;
+
+              setStageScale(newScale);
+            }}
             style={{
-              height: image && image?.height * stageScale,
-              width: image && image?.width * stageScale,
-              transform: `translate3d(${stagePos.x}px, ${stagePos.y}px, 0px)`,
+              transform: `scale(${stageScale})`,
+              transformOrigin: "top left",
             }}
-            onMouseMove={(event) => {
-              if (!isDrawMode && isDrag && !selectedRectId) {
-                event.preventDefault();
-                setStagePos({
-                  x: event.clientX - stageInitPos.x,
-                  y: event.clientY - stageInitPos.y,
-                });
-                setStageOffset(stagePos);
-              }
-            }}
-            onMouseUp={(event) => {
-              setIsDrag(false);
-              setStageInitPos(stagePos);
-            }}
-            onMouseLeave={(event) => {
-              setIsDrag(false);
-              setStageInitPos(stagePos);
-            }}
-            onMouseDown={(event) => {
-              setIsDrag(true);
-              setStageInitPos({
-                x: event.clientX - stageOffset.x,
-                y: event.clientY - stageOffset.y,
-              });
-            }}
-            onKeyDown={(e) => handleStageKeyDown(e)}
+            onContentMousedown={handleContentMousedown}
+            onContentMouseup={handleContentMouseup}
+            onContentMouseMove={handleContentMouseMove}
           >
-            <Stage
-              ref={stageRef}
-              width={image && image.width}
-              height={image && image.height}
-              onMouseDown={onStageClick}
-              onWheel={(e: Konva.KonvaEventObject<WheelEvent>) => {
-                e.evt.preventDefault();
-                const scaleBy = 1.04;
-                const newScale =
-                  e.evt.deltaY < 0
-                    ? stageScale * scaleBy
-                    : stageScale / scaleBy;
-                setStageScale(newScale);
-              }}
-              style={{
-                transform: `scale(${stageScale})`,
-                transformOrigin: "top left",
-              }}
-              onContentMousedown={handleContentMousedown}
-              onContentMouseup={handleContentMouseup}
-              onContentMouseMove={handleContentMouseMove}
-            >
-              <Layer>
-                <Image
-                  image={image}
-                  onMouseOver={(event) => {
-                    document.body.style.cursor = isDrawMode
-                      ? "crosshair"
-                      : "grab";
+            <Layer>
+              <Image
+                image={image}
+                onMouseOver={() => {
+                  document.body.style.cursor = isDrawMode
+                    ? "crosshair"
+                    : "grab";
+                }}
+                onMouseDown={() => {
+                  document.body.style.cursor = "grabbing";
+                }}
+                onMouseUp={() => {
+                  document.body.style.cursor = "grab";
+                }}
+                onMouseLeave={() => {
+                  document.body.style.cursor = "default";
+                }}
+              />
+              {ignoreAreas.map((rect, i) => (
+                <Rectangle
+                  key={rect.id}
+                  shapeProps={{
+                    x: rect.x,
+                    y: rect.y,
+                    width: rect.width,
+                    height: rect.height,
                   }}
-                  onMouseDown={(event) => {
-                    document.body.style.cursor = "grabbing";
-                  }}
-                  onMouseUp={(event) => {
-                    document.body.style.cursor = "grab";
-                  }}
-                  onMouseLeave={(event) => {
-                    document.body.style.cursor = "default";
+                  isSelected={rect.id === selectedRectId}
+                  onSelect={() => setSelectedRectId(rect.id)}
+                  onChange={(newAttrs: Konva.RectConfig) => {
+                    const rects = ignoreAreas.slice();
+
+                    rects[i].x = Math.round(newAttrs.x || 0);
+                    rects[i].y = Math.round(newAttrs.y || 0);
+                    rects[i].width = Math.round(
+                      newAttrs.width || MIN_RECT_SIDE_PIXEL,
+                    );
+                    rects[i].height = Math.round(
+                      newAttrs.height || MIN_RECT_SIDE_PIXEL,
+                    );
+
+                    setIgnoreAreas(rects);
                   }}
                 />
-                {ignoreAreas.map((rect, i) => {
-                  return (
-                    <Rectangle
-                      key={rect.id}
-                      shapeProps={{
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height,
-                      }}
-                      isSelected={rect.id === selectedRectId}
-                      onSelect={() => setSelectedRectId(rect.id)}
-                      onChange={(newAttrs: Konva.RectConfig) => {
-                        const rects = ignoreAreas.slice();
-
-                        rects[i].x = Math.round(newAttrs.x || 0);
-                        rects[i].y = Math.round(newAttrs.y || 0);
-                        rects[i].width = Math.round(
-                          newAttrs.width || MIN_RECT_SIDE_PIXEL,
-                        );
-                        rects[i].height = Math.round(
-                          newAttrs.height || MIN_RECT_SIDE_PIXEL,
-                        );
-
-                        setIgnoreAreas(rects);
-                      }}
-                    />
-                  );
-                })}
-                {tempIgnoreAreas.map((rect, i) => {
-                  return (
-                    <Rectangle
-                      key={i}
-                      shapeProps={{
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height,
-                      }}
-                    />
-                  );
-                })}
-              </Layer>
-            </Stage>
-          </div>
+              ))}
+              {tempIgnoreAreas.map((rect, i) => (
+                <Rectangle
+                  key={i}
+                  shapeProps={{
+                    x: rect.x,
+                    y: rect.y,
+                    width: rect.width,
+                    height: rect.height,
+                  }}
+                />
+              ))}
+            </Layer>
+          </Stage>
         </div>
-      )
+      </div>
     );
   };
 
   const imageLoading = () => {
     return (
-      imageStatus === "loading" && (
-        <Grid
-          container
-          direction="column"
-          alignItems="center"
-          justifyContent="center"
-          className={classes.progressContainer}
-        >
-          <Grid item>
-            <CircularProgress />
-          </Grid>
+      <Grid
+        container
+        direction="column"
+        alignItems="center"
+        justifyContent="center"
+        className={classes.progressContainer}
+      >
+        <Grid item>
+          <CircularProgress />
         </Grid>
-      )
+      </Grid>
     );
   };
 
   const imageFailed = () => {
-    return (!imageName || imageStatus === "failed") && <NoImagePlaceholder />;
+    // TODO: Use SVG with separate text for failing to load and why...
+    return <NoImagePlaceholder />;
   };
 
-  const imageCanvas = () => {
-    return (
-      <>
-        {imageFailed()}
-        {imageLoading()}
-        {imageLoaded()}
-      </>
-    );
-  };
-
-  return (
+  const imageCanvas = () => (
     <React.Fragment>
-      {imageName ? imageCanvas() : <NoImagePlaceholder />}
+      {(!imageName || imageStatus === "failed") && imageFailed()}
+      {imageStatus === "loading" && imageLoading()}
+      {imageStatus === "loaded" && imageLoaded()}
     </React.Fragment>
+  );
+
+  // TODO: Separate SVG with reason...
+  return (
+    <>
+      {imageName ? imageCanvas() : <NoImagePlaceholder />}
+    </>
   );
 };
