@@ -13,7 +13,7 @@ import {
   Tab,
   Select,
   MenuItem,
-} from "@material-ui/core";
+} from "@mui/material";
 import {
   useUserDispatch,
   useUserState,
@@ -66,7 +66,7 @@ const ProfilePage = () => {
     if (user && password) {
       usersService
         .changePassword(password)
-        .then((isChanged) => {
+        .then(() => {
           setPassword("");
         })
         .then(() =>
@@ -108,11 +108,10 @@ const ProfilePage = () => {
                           required
                           fullWidth
                           inputProps={{
-                            onChange: (event: any) =>
-                              setFirstName(
-                                (event.target as HTMLInputElement).value,
-                              ),
-                            "data-testId": "firstName",
+                            onChange: (
+                              event: React.ChangeEvent<HTMLInputElement>,
+                            ) => setFirstName(event.target.value),
+                            "data-testid": "firstName",
                           }}
                         />
                       </Grid>
@@ -129,11 +128,10 @@ const ProfilePage = () => {
                           required
                           fullWidth
                           inputProps={{
-                            onChange: (event: any) =>
-                              setLastName(
-                                (event.target as HTMLInputElement).value,
-                              ),
-                            "data-testId": "lastName",
+                            onChange: (
+                              event: React.ChangeEvent<HTMLInputElement>,
+                            ) => setLastName(event.target.value),
+                            "data-testid": "lastName",
                           }}
                         />
                       </Grid>
@@ -150,16 +148,16 @@ const ProfilePage = () => {
                           required
                           fullWidth
                           inputProps={{
-                            onChange: (event: any) =>
-                              setEmail(
-                                (event.target as HTMLInputElement).value,
-                              ),
-                            "data-testId": "email",
+                            onChange: (
+                              event: React.ChangeEvent<HTMLInputElement>,
+                            ) => setEmail(event.target.value),
+                            "data-testid": "email",
                           }}
                         />
                       </Grid>
                       <Grid item xs={12}>
                         <Select
+                          variant="standard"
                           id="role"
                           labelId="role"
                           displayEmpty
@@ -183,7 +181,7 @@ const ProfilePage = () => {
                           type="submit"
                           color="primary"
                           variant="outlined"
-                          data-testId="submit"
+                          data-testid="submit"
                         >
                           Update
                         </Button>
@@ -215,11 +213,10 @@ const ProfilePage = () => {
                           required
                           fullWidth
                           inputProps={{
-                            onChange: (event: any) =>
-                              setPassword(
-                                (event.target as HTMLInputElement).value,
-                              ),
-                            "data-testId": "password",
+                            onChange: (
+                              event: React.ChangeEvent<HTMLInputElement>,
+                            ) => setPassword(event.target.value),
+                            "data-testid": "password",
                           }}
                         />
                       </Grid>
@@ -232,7 +229,7 @@ const ProfilePage = () => {
                           type="submit"
                           color="primary"
                           variant="outlined"
-                          data-testId="submit"
+                          data-testid="submit"
                         >
                           Update
                         </Button>
@@ -267,7 +264,7 @@ const ProfilePage = () => {
                     indicatorColor="primary"
                     textColor="primary"
                     onChange={(
-                      event: React.ChangeEvent<{}>,
+                      event: React.ChangeEvent<unknown>,
                       newValue: number,
                     ) => {
                       setTabIndex(newValue);
